test_aimassist.py
import unittest
from lib import aimassist
import unittest
import torch
from lib import QuantumNeuralNet
import unittest
import torch
from lib.aimassist import QuantumNeuralNet

class TestAimAssist(unittest.TestCase):
    def setUp(self):
        self.aim_assist = aimassist(sensitivity=1.5)

    def test_move_cursor(self):
        # Example test for cursor movement (you may need to mock pyautogui for proper testing)
        self.aim_assist.move_cursor(100, 200)
        # Assert the expected behavior, depending on how you choose to test it

    def test_click_target(self):
        # Example test for clicking (again, mocking may be required)
        self.aim_assist.click_target()
        # Assert the expected behavior
        class TestQuantumNeuralNet(unittest.TestCase):
            def setUp(self):
                self.qnn = QuantumNeuralNet()

            def test_forward(self):
                # Create a random input tensor
                x = torch.randn(1, 4)

                # Perform forward pass
                output = self.qnn.forward(x)

                # Assert the output shape and type
                self.assertEqual(output.shape, torch.Size([1, 2]))
                self.assertIsInstance(output, torch.Tensor)

            def test_train_model(self):
                # Create a dummy train loader
                train_loader = [(torch.randn(1, 4), torch.tensor([0]))]

                # Train the model
                self.qnn.train_model(train_loader, num_epochs=1, learning_rate=0.001)

                # Assert that the model parameters have been updated
                self.assertTrue(any(p.requires_grad for p in self.qnn.parameters()))

            def test_evaluate_model(self):
                # Create a dummy test loader
                test_loader = [(torch.randn(1, 4), torch.tensor([0]))]

                # Evaluate the model
                self.qnn.evaluate_model(test_loader)

                # Assert that the model is in evaluation mode
                self.assertFalse(self.qnn.training)

        if __name__ == "__main__":
            unittest.main()
if __name__ == "__main__":
    class TestQuantumNeuralNet(unittest.TestCase):
        def setUp(self):
            self.qnn = QuantumNeuralNet()

        def test_forward(self):
            # Create a random input tensor
            x = torch.randn(1, 4)

            # Perform forward pass
            output = self.qnn.forward(x)

            # Assert the output shape and type
            self.assertEqual(output.shape, torch.Size([1, 2]))
            self.assertIsInstance(output, torch.Tensor)

        def test_train_model(self):
            # Create a dummy train loader
            train_loader = [(torch.randn(1, 4), torch.tensor([0]))]

            # Train the model
            self.qnn.train_model(train_loader, num_epochs=1, learning_rate=0.001)

            # Assert that the model parameters have been updated
            self.assertTrue(any(p.requires_grad for p in self.qnn.parameters()))

        def test_evaluate_model(self):
            # Create a dummy test loader
            test_loader = [(torch.randn(1, 4), torch.tensor([0]))]

            # Evaluate the model
            self.qnn.evaluate_model(test_loader)

            # Assert that the model is in evaluation mode
            self.assertFalse(self.qnn.training)

    if __name__ == "__main__":
        unittest.main()
    unittest.main()
